% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/displayTiles.R
\name{displayTiles}
\alias{displayTiles}
\title{Display image frames in a row}
\usage{
displayTiles(img, lwd = 10)
}
\arguments{
\item{img}{an \code{Image}}

\item{lwd}{width of the grid lines between tiled images}
}
\description{
Given a seqence of frames, generate a single image with frames tiled in a row.
}

